V3.0
Fixed:
Script crashing on the input of an incorrect private key
Script allows also public keys to check whether there are Solanas or not
Script can donate a percentage to the creator of the script. if users enables it in the config.
Can now check if there are any other tokens than Solana if it is from the solana network (for example memecoins or staked solana)